dev-qa-icl: Fix small bugs

+++ QuestionAnswering/QA_ICL/backend/fastapi_app/model/kg/ontokin.py
@@ -50,7 +50,7 @@ class OntokinTransportModel(RDFEntity):
         path=ONTOKIN.hasLJPotentialWellDepth
     )
     polarizability: OntokinHasValueHasUnit = RDFField(path=ONTOKIN.hasPolarizability)
-    rotational_relaxation_collision_number: int = RDFField(
+    rotational_relaxation_collision_number: float = RDFField(
         path=ONTOKIN.hasRotationalRelaxationCollisionNumber / ONTOKIN.value
     )
     shape_index: int = RDFField(path=ONTOKIN.hasShapeIndex / ONTOKIN.value)
